requirements.txt
jax
flax
torch
pytorch-lightning
wandb
pyyaml
